Move circuit construction to separate module
