+++ mthree/mitigation.py
@@ -22,92 +22,17 @@ import numpy as np
 import scipy.linalg as la
 import scipy.sparse.linalg as spla
 import orjson
-from qiskit import QuantumCircuit, transpile, execute
+from qiskit import transpile, execute
 from qiskit.providers import BaseBackend
 
+from mthree.circuits import (_tensor_meas_states, _marg_meas_states,
+                             balanced_cal_strings, balanced_cal_circuits)
 from mthree.matrix import _reduced_cal_matrix, sdd_check
 from mthree.utils import counts_to_vector, vector_to_quasiprobs
 from mthree.norms import ainv_onenorm_est_lu, ainv_onenorm_est_iter
 from mthree.matvec import M3MatVec
 from mthree.exceptions import M3Error
 from mthree.classes import QuasiCollection
-
-
-def _tensor_meas_states(qubit, num_qubits, initial_reset=False):
-    """Construct |0> and |1> states
-    for independent 1Q cals.
-    """
-    qc0 = QuantumCircuit(num_qubits, 1)
-    if initial_reset:
-        qc0.reset(qubit)
-    qc0.measure(qubit, 0)
-    qc1 = QuantumCircuit(num_qubits, 1)
-    if initial_reset:
-        qc1.reset(qubit)
-    qc1.x(qubit)
-    qc1.measure(qubit, 0)
-    return [qc0, qc1]
-
-
-def _marg_meas_states(num_qubits, initial_reset=False):
-    """Construct all zeros and all ones states
-    for marginal 1Q cals.
-    """
-    qc0 = QuantumCircuit(num_qubits)
-    if initial_reset:
-        qc0.reset(range(num_qubits))
-    qc0.measure_all()
-    qc1 = QuantumCircuit(num_qubits)
-    if initial_reset:
-        qc1.reset(range(num_qubits))
-    qc1.x(range(num_qubits))
-    qc1.measure_all()
-    return [qc0, qc1]
-
-
-def _balanced_cal_strings(num_qubits):
-    """Compute the 2*num_qubits strings for balanced calibration.
-
-    Parameters:
-        num_qubits (int): Number of qubits to be measured.
-
-    Returns:
-        list: List of strings for balanced calibration circuits.
-    """
-    strings = []
-    for rep in range(1, num_qubits+1):
-        str1 = ''
-        str2 = ''
-        for jj in range(int(np.ceil(num_qubits / rep))):
-            str1 += str(jj % 2) * rep
-            str2 += str((jj+1) % 2) * rep
-
-        strings.append(str1[:num_qubits])
-        strings.append(str2[:num_qubits])
-    return strings
-
-
-def _balanced_cal_circuits(cal_strings, initial_reset=False):
-    """Build balanced calibration circuits.
-
-    Parameters:
-        cal_strings (list): List of strings for balanced cal circuits.
-
-    Returns:
-        list: List of balanced cal circuits.
-    """
-    num_qubits = len(cal_strings[0])
-    circs = []
-    for string in cal_strings:
-        qc = QuantumCircuit(num_qubits)
-        if initial_reset:
-            qc.reset(range(num_qubits))
-        for idx, bit in enumerate(string[::-1]):
-            if bit == '1':
-                qc.x(idx)
-        qc.measure_all()
-        circs.append(qc)
-    return circs
 
 
 class M3Mitigation():
@@ -362,8 +287,8 @@ class M3Mitigation():
             trans_qcs = transpile(circs, self.system,
                                   initial_layout=qubits, optimization_level=0)
         elif method == 'balanced':
-            cal_strings = _balanced_cal_strings(num_cal_qubits)
-            circs = _balanced_cal_circuits(cal_strings, initial_reset=initial_reset)
+            cal_strings = balanced_cal_strings(num_cal_qubits)
+            circs = balanced_cal_circuits(cal_strings, initial_reset=initial_reset)
             trans_qcs = transpile(circs, self.system,
                                   initial_layout=qubits, optimization_level=0)
         # Indeopendent
